modify view of computing power

+++ oas-app/oases/ComputingPower/src/main/java/com/oases/computingpower/presenter/view/ComputingPowerMainView.kt
@@ -4,9 +4,10 @@ import com.oases.base.presenter.view.BaseView
 import com.oases.computingpower.data.protocol.MailStatusResp
 import com.oases.computingpower.data.protocol.PhoneStatusResp
 import com.oases.computingpower.data.protocol.PowerActivityStatusResp
+import java.math.BigDecimal
 
 interface ComputingPowerMainView: BaseView {
-    fun onGetPower(power: Int)
+    fun onGetPower(power: BigDecimal)
     fun onCheckPhoneResult(result:PhoneStatusResp)
     fun onCheckMailResult(result: MailStatusResp)
     fun onInquirePowerActivityStatusResult(result: PowerActivityStatusResp)
